chore: Move source files to correct directory
Forgot to move these after swapping the namespace out
