+++ src/main/kotlin/io/github/atriusx/util/CompositesUtil.kt
@@ -1,4 +1,4 @@
-package xyz.atrius.util
+package io.github.atriusx.util
 
 import org.gradle.api.Project
 import org.gradle.api.initialization.Settings
